Moved commands and reply modules into the server module.
