+++ src/lib.rs
@@ -25,9 +25,6 @@
 pub mod server;
 pub use crate::server::Server;
 
-pub(crate) mod commands;
-pub(crate) mod reply;
-
 /// Contains the `Authenticator` trait that is used by the `Server` to authenticate users, as well
 /// as its various implementations.
 pub mod auth;
